curiousitydriven/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
ca-certificates=2021.4.13=haa95532_1
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py38haa95532_0
chardet=4.0.0=pypi_0
cloudpickle=1.6.0=pypi_0
cudatoolkit=10.2.89=h74a9793_1
flatbuffers=1.12=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
google-auth=1.31.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.1=pypi_0
gym=0.18.3=pypi_0
gym-super-mario-bros=7.3.2=pypi_0
h5py=3.1.0=pypi_0
idna=2.10=pypi_0
intel-openmp=2021.2.0=haa95532_616
jpeg=9b=hb83a4c4_2
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.4=pypi_0
mkl=2021.2.0=haa95532_296
mkl-service=2.3.0=py38h2bbff1b_1
mkl_fft=1.3.0=py38h277e83a_2
mkl_random=1.2.1=py38hf11a4ad_2
nes-py=8.1.6=pypi_0
ninja=1.10.2=h6d14046_1
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
openssl=1.1.1k=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
pillow=8.2.0=py38h4fa10fc_0
pip=21.1.1=py38haa95532_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyglet=1.5.0=pypi_0
python=3.8.10=hdbf39b2_7
pytorch=1.8.1=py3.8_cuda10.2_cudnn7_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.6.3=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
sqlite=3.35.4=h2bbff1b_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.10=he774522_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cu102
tqdm=4.61.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.5=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
